refactor: move popups to sub folder
