+++ packages/desktop/components/popup/Popup.svelte
@@ -8,51 +8,51 @@
     import { clickOutside } from '@core/utils/ui'
 
     // Popups
-    import AccountSwitcherPopup from './AccountSwitcherPopup.svelte'
-    import ActivityDetailsPopup from './ActivityDetailsPopup.svelte'
-    import AddNodePopup from './AddNodePopup.svelte'
-    import AddProposalPopup from './AddProposalPopup.svelte'
-    import AliasConfirmationPopup from './AliasConfirmationPopup.svelte'
-    import BackupStrongholdPopup from './BackupStrongholdPopup.svelte'
-    import BurnNativeTokensPopup from './BurnNativeTokensPopup.svelte'
-    import BurnNativeTokensConfirmationPopup from './BurnNativeTokensConfirmationPopup.svelte'
-    import ConfirmationPopup from './ConfirmationPopup.svelte'
-    import ConnectLedgerPopup from './ConnectLedgerPopup.svelte'
-    import CreateAccountPopup from './CreateAccountPopup.svelte'
-    import DeepLinkErrorPopup from './DeepLinkErrorPopup.svelte'
-    import DeleteAccountPopup from './DeleteAccountPopup.svelte'
-    import DeleteProfilePopup from './DeleteProfilePopup.svelte'
-    import DiagnosticsPopup from './DiagnosticsPopup.svelte'
-    import EnableLedgerBlindSigningPopup from './EnableLedgerBlindSigningPopup.svelte'
-    import ErrorLogPopup from './ErrorLogPopup.svelte'
-    import FaucetRequestPopup from './FaucetRequestPopup.svelte'
-    import ImportErc20TokenFormPopup from './ImportErc20TokenFormPopup.svelte'
-    import LedgerAppGuidePopup from './LedgerAppGuidePopup.svelte'
-    import LedgerConnectionGuidePopup from './LedgerConnectionGuidePopup.svelte'
-    import LegalUpdatePopup from './LegalUpdatePopup.svelte'
-    import ManageAccountPopup from './ManageAccountPopup.svelte'
-    import ManageVotingPowerPopup from './ManageVotingPowerPopup.svelte'
-    import MintNativeTokenConfirmationPopup from './MintNativeTokenConfirmationPopup.svelte'
-    import MintNativeTokenFormPopup from './MintNativeTokenFormPopup.svelte'
-    import MintNftConfirmationPopup from './MintNftConfirmationPopup.svelte'
-    import MintNftFormPopup from './MintNftFormPopup.svelte'
-    import NodeAuthRequiredPopup from './NodeAuthRequiredPopup.svelte'
-    import NodeInfoPopup from './NodeInfoPopup.svelte'
-    import ReceiveAddressPopup from './ReceiveAddressPopup.svelte'
-    import RemoveProposalPopup from './RemoveProposalPopup.svelte'
-    import RevotePopup from './RevotePopup.svelte'
-    import SignMessagePopup from './SignMessagePopup.svelte'
-    import SendFlowPopup from './SendFlowPopup.svelte'
-    import StopVotingPopup from './StopVotingPopup.svelte'
-    import BalanceBreakdownPopup from './BalanceBreakdownPopup.svelte'
-    import TestDeepLinkFormPopup from './TestDeepLinkFormPopup.svelte'
-    import TokenInformationPopup from './TokenInformationPopup.svelte'
-    import UnlockStrongholdPopup from './UnlockStrongholdPopup.svelte'
-    import VerifyLedgerTransactionPopup from './VerifyLedgerTransactionPopup.svelte'
-    import CheckForUpdatesPopup from './CheckForUpdatesPopup.svelte'
-    import VoteForProposal from './VoteForProposalPopup.svelte'
-    import VotingPowerToZeroPopup from './VotingPowerToZeroPopup.svelte'
-    import WalletFinderPopup from './WalletFinderPopup.svelte'
+    import AccountSwitcherPopup from './popups/AccountSwitcherPopup.svelte'
+    import ActivityDetailsPopup from './popups/ActivityDetailsPopup.svelte'
+    import AddNodePopup from './popups/AddNodePopup.svelte'
+    import AddProposalPopup from './popups/AddProposalPopup.svelte'
+    import AliasConfirmationPopup from './popups/AliasConfirmationPopup.svelte'
+    import BackupStrongholdPopup from './popups/BackupStrongholdPopup.svelte'
+    import BurnNativeTokensPopup from './popups/BurnNativeTokensPopup.svelte'
+    import BurnNativeTokensConfirmationPopup from './popups/BurnNativeTokensConfirmationPopup.svelte'
+    import ConfirmationPopup from './popups/ConfirmationPopup.svelte'
+    import ConnectLedgerPopup from './popups/ConnectLedgerPopup.svelte'
+    import CreateAccountPopup from './popups/CreateAccountPopup.svelte'
+    import DeepLinkErrorPopup from './popups/DeepLinkErrorPopup.svelte'
+    import DeleteAccountPopup from './popups/DeleteAccountPopup.svelte'
+    import DeleteProfilePopup from './popups/DeleteProfilePopup.svelte'
+    import DiagnosticsPopup from './popups/DiagnosticsPopup.svelte'
+    import EnableLedgerBlindSigningPopup from './popups/EnableLedgerBlindSigningPopup.svelte'
+    import ErrorLogPopup from './popups/ErrorLogPopup.svelte'
+    import FaucetRequestPopup from './popups/FaucetRequestPopup.svelte'
+    import ImportErc20TokenFormPopup from './popups/ImportErc20TokenFormPopup.svelte'
+    import LedgerAppGuidePopup from './popups/LedgerAppGuidePopup.svelte'
+    import LedgerConnectionGuidePopup from './popups/LedgerConnectionGuidePopup.svelte'
+    import LegalUpdatePopup from './popups/LegalUpdatePopup.svelte'
+    import ManageAccountPopup from './popups/ManageAccountPopup.svelte'
+    import ManageVotingPowerPopup from './popups/ManageVotingPowerPopup.svelte'
+    import MintNativeTokenConfirmationPopup from './popups/MintNativeTokenConfirmationPopup.svelte'
+    import MintNativeTokenFormPopup from './popups/MintNativeTokenFormPopup.svelte'
+    import MintNftConfirmationPopup from './popups/MintNftConfirmationPopup.svelte'
+    import MintNftFormPopup from './popups/MintNftFormPopup.svelte'
+    import NodeAuthRequiredPopup from './popups/NodeAuthRequiredPopup.svelte'
+    import NodeInfoPopup from './popups/NodeInfoPopup.svelte'
+    import ReceiveAddressPopup from './popups/ReceiveAddressPopup.svelte'
+    import RemoveProposalPopup from './popups/RemoveProposalPopup.svelte'
+    import RevotePopup from './popups/RevotePopup.svelte'
+    import SignMessagePopup from './popups/SignMessagePopup.svelte'
+    import SendFlowPopup from './popups/SendFlowPopup.svelte'
+    import StopVotingPopup from './popups/StopVotingPopup.svelte'
+    import BalanceBreakdownPopup from './popups/BalanceBreakdownPopup.svelte'
+    import TestDeepLinkFormPopup from './popups/TestDeepLinkFormPopup.svelte'
+    import TokenInformationPopup from './popups/TokenInformationPopup.svelte'
+    import UnlockStrongholdPopup from './popups/UnlockStrongholdPopup.svelte'
+    import VerifyLedgerTransactionPopup from './popups/VerifyLedgerTransactionPopup.svelte'
+    import CheckForUpdatesPopup from './popups/CheckForUpdatesPopup.svelte'
+    import VoteForProposal from './popups/VoteForProposalPopup.svelte'
+    import VotingPowerToZeroPopup from './popups/VotingPowerToZeroPopup.svelte'
+    import WalletFinderPopup from './popups/WalletFinderPopup.svelte'
 
     export let id: PopupId
     export let props: any
