+++ packages/desktop/components/popup/popups/ActivityDetailsPopup.svelte
