+++ packages/desktop/components/popup/popups/BackupStrongholdPopup.svelte
