+++ packages/desktop/components/popup/popups/BalanceBreakdownPopup.svelte
