+++ packages/desktop/components/popup/popups/BurnNativeTokensConfirmationPopup.svelte
