+++ packages/desktop/components/popup/popups/BurnNativeTokensPopup.svelte
