+++ packages/desktop/components/popup/popups/DiagnosticsPopup.svelte
