+++ packages/desktop/components/popup/popups/EnableLedgerBlindSigningPopup.svelte
