+++ packages/desktop/components/popup/popups/ErrorLogPopup.svelte
