+++ packages/desktop/components/popup/popups/LedgerAppGuidePopup.svelte
