+++ packages/desktop/components/popup/popups/LedgerConnectionGuidePopup.svelte
