+++ packages/desktop/components/popup/popups/LegalUpdatePopup.svelte
