+++ packages/desktop/components/popup/popups/ManageVotingPowerPopup.svelte
