+++ packages/desktop/components/popup/popups/MintNativeTokenConfirmationPopup.svelte
