+++ packages/desktop/components/popup/popups/MintNftConfirmationPopup.svelte
