+++ packages/desktop/components/popup/popups/MintNftFormPopup.svelte
