+++ packages/desktop/components/popup/popups/NodeAuthRequiredPopup.svelte
