+++ packages/desktop/components/popup/popups/NodeInfoPopup.svelte
