+++ packages/desktop/components/popup/popups/RemoveProposalPopup.svelte
