+++ packages/desktop/components/popup/popups/StopVotingPopup.svelte
