+++ packages/desktop/components/popup/popups/TokenInformationPopup.svelte
