+++ packages/desktop/components/popup/popups/UnlockStrongholdPopup.svelte
