+++ packages/desktop/components/popup/popups/VoteForProposalPopup.svelte
