+++ packages/desktop/components/popup/popups/VotingPowerToZeroPopup.svelte
