+++ packages/desktop/components/popup/popups/WalletFinderPopup.svelte
